frequently used func transfered to separate file
